Remove itctools2 language, combine directories
