+++ MANIFEST.in
@@ -2,5 +2,5 @@ include LICENSE
 include MANIFEST.in
 include versioneer.py
 
-graft itctools2
+graft itctools
 global-exclude *.py[cod] __pycache__ *.so
