+++ appveyor.yml
@@ -90,7 +90,7 @@ install:
 build: false
 
 test_script:
-  - pytest -v --cov=itctools2 itctools2\\tests
+  - pytest -v --cov=itctools itctools\\tests
 
 on_success:
   - codecov
