+++ docs/api.rst
@@ -4,4 +4,4 @@ API Documentation
 .. autosummary::
    :toctree: autosummary
 
-   itctools2.canvas
+   itctools.canvas
