+++ docs/autosummary/itctools.canvas.rst
@@ -1,0 +1,6 @@
+itctools.canvas
+================
+
+.. currentmodule:: itctools
+
+.. autofunction:: canvas
