+++ docs/conf.py
@@ -17,12 +17,12 @@ import os
 import sys
 sys.path.insert(0, os.path.abspath('..'))
 
-import itctools2
+import itctools
 
 
 # -- Project information -----------------------------------------------------
 
-project = 'itctools2'
+project = 'itctools'
 copyright = ("2020, Chodera Lab. Project structure based on the "
              "Computational Molecular Science Python Cookiecutter version 1.1")
 author = 'Chodera Lab'
@@ -117,7 +117,7 @@ html_static_path = ['_static']
 # -- Options for HTMLHelp output ---------------------------------------------
 
 # Output file base name for HTML help builder.
-htmlhelp_basename = 'itctools2doc'
+htmlhelp_basename = 'itctoolsdoc'
 
 
 # -- Options for LaTeX output ------------------------------------------------
@@ -144,8 +144,8 @@ latex_elements = {
 # (source start file, target name, title,
 #  author, documentclass [howto, manual, or own class]).
 latex_documents = [
-    (master_doc, 'itctools2.tex', 'itctools2 Documentation',
-     'itctools2', 'manual'),
+    (master_doc, 'itctools.tex', 'itctools Documentation',
+     'itctools', 'manual'),
 ]
 
 
@@ -154,7 +154,7 @@ latex_documents = [
 # One entry per manual page. List of tuples
 # (source start file, name, description, authors, manual section).
 man_pages = [
-    (master_doc, 'itctools2', 'itctools2 Documentation',
+    (master_doc, 'itctools', 'itctools Documentation',
      [author], 1)
 ]
 
@@ -165,8 +165,8 @@ man_pages = [
 # (source start file, target name, title, author,
 #  dir menu entry, description, category)
 texinfo_documents = [
-    (master_doc, 'itctools2', 'itctools2 Documentation',
-     author, 'itctools2', 'An itctools repo based on the molssi cookiecutter',
+    (master_doc, 'itctools', 'itctools Documentation',
+     author, 'itctools', 'An itctools repo based on the molssi cookiecutter',
      'Miscellaneous'),
 ]
