+++ examples/host_guest/host-guest-itc-Rm.txt
@@ -1,0 +1,15 @@
+#, Optimal, Actual
+1, 1.11433, 1.09769
+2, 1.21750, 1.19931
+3, 3.90871, 3.85031
+4, 1.84497, 1.81741
+5, 1.77450, 1.74799
+6, 2.21184, 2.17880
+7, 1.10000, 1.08357
+8, 1.10000, 1.08357
+9, 1.10000, 1.08357
+10, 2.27045, 2.23653
+11, 1.10000, 1.08357
+12, 1.10000, 1.08357
+13, 1.10000, 1.08357
+14, 1.10000, 1.08357
