+++ itctools/_version.py
@@ -43,7 +43,7 @@ def get_config():
     cfg.style = "pep440"
     cfg.tag_prefix = ""
     cfg.parentdir_prefix = "None"
-    cfg.versionfile_source = "itctools2/_version.py"
+    cfg.versionfile_source = "itctools/_version.py"
     cfg.verbose = False
     return cfg
