+++ itctools/itctools.py
@@ -1,3 +1,10 @@
+"""
+itctools.py
+An itctools repo based on the molssi cookiecutter
+
+Handles the primary functions
+"""
+
 import itertools
 from pint import UnitRegistry
 import numpy
@@ -26,3 +33,31 @@ def compute_rm(c):
     """
     rm = 6.4 / numpy.power(c, 0.2) + 13 / c
     return rm
+
+
+def canvas(with_attribution=True):
+    """
+    Placeholder function to show example docstring (NumPy format)
+
+    Replace this function and doc string for your own project
+
+    Parameters
+    ----------
+    with_attribution : bool, Optional, default: True
+        Set whether or not to display who the quote is from
+
+    Returns
+    -------
+    quote : str
+        Compiled string including quote and optional attribution
+    """
+
+    quote = "The code is but a canvas to our imagination."
+    if with_attribution:
+        quote += "\n\t- Adapted from Henry David Thoreau"
+    return quote
+
+
+if __name__ == "__main__":
+    # Do something if this file is invoked on its own
+    print(canvas())
