--- itctools2/tests/__init__.py
@@ -1,3 +1,0 @@
-"""
-Empty init file in case you choose a package besides PyTest such as Nose which may look for such a file
-"""
