--- itctools2/tests/test_itctools2.py
@@ -1,12 +1,0 @@
-"""
-Unit and regression test for the itctools2 package.
-"""
-
-# Import package, test suite, and other packages as needed
-import itctools2
-import pytest
-import sys
-
-def test_itctools2_imported():
-    """Sample test, will always pass so long as import statement worked"""
-    assert "itctools2" in sys.modules
