+++ setup.cfg
@@ -11,7 +11,7 @@ omit =
     # Omit the tests
     */tests/*
     # Omit generated versioneer
-    itctools2/_version.py
+    itctools/_version.py
 
 [yapf]
 # YAPF, in .style.yapf files this shows up as "[style]" header
@@ -27,8 +27,8 @@ max-line-length = 119
 # Automatic version numbering scheme
 VCS = git
 style = pep440
-versionfile_source = itctools2/_version.py
-versionfile_build = itctools2/_version.py
+versionfile_source = itctools/_version.py
+versionfile_build = itctools/_version.py
 tag_prefix = ''
 
 [aliases]
